Separates the core loops to functions
It makes sense to skip the parallel loop when there is just a single
element to loop over, since the threading overhead is actually
significant otherwise. This meant that the core loops were duplicated
in the functions, which is very error-prone.
Now, the core implementations are in separated c-functions called from
the python interface layer. These are implemented with the element
indexing taking place inside the core functions. The alternative is to
pass slices into the core functions, which seems to be slower for some
reason.

+++ shetar/_bases.pxd
@@ -1,4 +1,4 @@
 
-cdef int associated_legendre_index(int order, int mode) nogil
+cdef Py_ssize_t associated_legendre_index(Py_ssize_t order, Py_ssize_t mode) nogil
 
-cdef int spherical_expansion_index(int order, int mode) nogil
+cdef Py_ssize_t spherical_expansion_index(Py_ssize_t order, Py_ssize_t mode) nogil
